Fix Response Headers to pass the tests
There is a test for each of the three post requests (contract call, get balance, transfer) which expects `'Content-Type'` to be `'application/json'` on the response header. (The test description is misleading though, because is refers to the request header)

Express automatically adds `'charset=UTF-8'`, which makes the test fail.

The changes in this PR fix the issue.